bandit

+++ train_bandit.lua
@@ -61,7 +61,7 @@ end
 
 
 local function sample_action(model_output)
-    local result = torch.multinomial(model_output,1)
+    local result = torch.multinomial(torch.LongTensor(model_output),1)
     return result
 end
